requirements.txt
streamlit
pandas
openpyxl
gtts
fpdf
google-generativeai